requirements/base.txt
requests==2.24.0
click===7.1.2
gevent==20.6.2
PyYAML==5.3.1